+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## Version develop
+- [Fix: Add `IJavaScriptSnippet` service interface and update the `IServiceCollection` extension to register it for `JavaScriptSnippet`.](https://github.com/microsoft/ApplicationInsights-aspnetcore/issues/890)
+
 ## Version 2.7.0
- - Updated Web/Base SDK version dependency to 2.10.0
+- Updated Web/Base SDK version dependency to 2.10.0
 - [Remove unused reference to System.Net.Http](https://github.com/microsoft/ApplicationInsights-aspnetcore/pull/879)
 
 ## Version 2.7.0-beta4